chore: don't rely on getting old auctioneer instance
The vaults upgrade consumed instance.auctioneer and compared it to the
new instance, passed from the new auctioneer proposal in a single-use
promise, and failed if they weren't different. In this software
upgrade, the instance.auctioneer value might be the new value, so this
test can fail. Since the single-use promise comes directly from the
new auction proposal, if it has a value, it's the new one.